refactor(boundaries): extract persona_drift leaf module (refs #1367)
Break the ``heartbeats.local`` lazy-import of ``pollypm.supervisor``.
``detect_persona_drift`` (plus its ``_ROLE_PERSONA_MARKER`` table and
``_IDENTITY_CLAIM_PATTERNS`` phrasings) moves to a new leaf module
``pollypm.persona_drift`` that depends on nothing else in the package.
``heartbeats.local`` top-imports the predicate directly; ``supervisor``
keeps the names as re-exports for back-compat with tests and any
external callers.

The lazy import in ``heartbeats/local._check_persona_drift`` (the
``try: from pollypm.supervisor import detect_persona_drift`` block)
was the documented #1367 cycle-breaker for heartbeats ↔ supervisor.
Pure module-boundary move — no behavior change, no callsite signature
change. Verified by import-smoke + `_ROLE_PERSONA_MARKER` identity
through the supervisor re-export.
