+++ src/pollypm/supervisor.py
@@ -76,6 +76,15 @@ from pollypm.heartbeats.api import SupervisorHeartbeatAPI
 from pollypm.knowledge_extract import EXTRACTION_INTERVAL_SECONDS
 from pollypm.models import AccountConfig, ProviderKind, SessionConfig, SessionLaunchSpec
 from pollypm.onboarding import _prime_claude_home
+# Canonical persona-drift definitions live in :mod:`pollypm.persona_drift`
+# so ``heartbeats.local`` can import :func:`detect_persona_drift` without
+# lazy-importing supervisor (refs #1367). Re-exported below for
+# back-compat with existing callers and tests.
+from pollypm.persona_drift import (  # noqa: F401  (re-exported)
+    _IDENTITY_CLAIM_PATTERNS,
+    _ROLE_PERSONA_MARKER,
+    detect_persona_drift,
+)
 from pollypm.projects import ensure_project_scaffold
 from pollypm.projects import project_checkpoints_dir, project_transcripts_dir, project_worktrees_dir, release_session_lock
 from pollypm.providers.claude.resume import recorded_session_id as _recorded_claude_session_id
@@ -161,13 +170,10 @@ _OWNER_PREFIXES = {
 # (launch, target) tuple crosses where one role's control prompt
 # lands in a different role's window. ``worker`` has no stable
 # persona and ``triage`` does not currently brand itself, so both
-# are intentionally omitted.
-_ROLE_PERSONA_MARKER: dict[str, str] = {
-    "operator-pm": "Polly",
-    "reviewer": "Russell",
-    "architect": "Archie",
-    "heartbeat-supervisor": "Heartbeat",
-}
+# are intentionally omitted. Canonical definitions (the marker table,
+# identity-claim phrasings, and the :func:`detect_persona_drift`
+# predicate) live in :mod:`pollypm.persona_drift` — imported at the
+# top of this module and re-exported for back-compat.
 
 
 # Identity preamble prepended to recovery prompts so the agent does not
@@ -243,63 +249,6 @@ def _identity_preamble_for_role(role: str | None) -> str:
     if not role:
         return ""
     return _ROLE_IDENTITY_PREAMBLE.get(role, "")
-
-
-# Phrasings that unambiguously claim a persona identity — e.g.
-# "Standing by as Russell", "I'm Polly", "Holding as Russell". Used
-# by :func:`detect_persona_drift` to decide whether the pane shows
-# a DIFFERENT persona's identity claim, which is the symptom of a
-# mid-flight persona swap (#757). Neutral mentions ("let me notify
-# Polly") don't match these patterns.
-_IDENTITY_CLAIM_PATTERNS: tuple[str, ...] = (
-    "standing by as {marker}",
-    "holding as {marker}",
-    "i am {marker}",
-    "i'm {marker}",
-    "continuing as {marker}",
-    "acting as {marker}",
-    "as {marker}, ",
-    "as {marker}.",
-    "as {marker},",
-    "initialized as {marker}",
-)
-
-
-def detect_persona_drift(role: str, pane_text: str) -> str | None:
-    """Return the name of the drifted-to persona, or None on no drift.
-
-    Looks for strong identity-claim phrasings that reference a role
-    OTHER than the session's configured role. Conservative by design:
-    casual mentions of another persona name don't trip the detector —
-    only phrasings the session would use to assert its own identity.
-
-    Used by the heartbeat (#757) to catch sessions that started with
-    the correct role but drifted mid-flight — either through kickoff
-    clobber (#758) or prompt-injection (#755). Kickoff-time drift is
-    caught earlier by :func:`Supervisor._assert_session_launch_matches`.
-
-    Returns the detected persona name (e.g. ``"Russell"``) so callers
-    can surface it in the alert message, or ``None`` when no drift is
-    observed.
-    """
-    if not pane_text or not role:
-        return None
-    expected = _ROLE_PERSONA_MARKER.get(role)
-    lowered = pane_text.lower()
-    for other_role, other_marker in _ROLE_PERSONA_MARKER.items():
-        if other_role == role:
-            continue
-        # Avoid false positives when the expected marker also appears
-        # in the pane — both present means the session is legitimately
-        # discussing multiple personas (e.g. Polly reviewing Russell's
-        # output).
-        if expected and expected.lower() in lowered:
-            continue
-        for pattern in _IDENTITY_CLAIM_PATTERNS:
-            needle = pattern.format(marker=other_marker.lower())
-            if needle in lowered:
-                return other_marker
-    return None
 
 
 def _prefix_for_owner(owner: str, text: str) -> str:
